+++ lib/appium_lib_core/common/base/bridge/msjsonwp.rb
@@ -1,0 +1,13 @@
+module Appium
+  module Core
+    class Base
+      class Bridge
+        class MJSONWP < ::Selenium::WebDriver::Remote::OSS::Bridge
+          def commands(command)
+            ::Appium::Core::Commands::COMMANDS_EXTEND_MJSONWP[command]
+          end
+        end # class MJSONWP
+      end # class Bridge
+    end # class Base
+  end # module Core
+end # module Appium
